reorganized data mapping files
