+++ src/js/components/aboutTheData/AboutTheDataModal.jsx
@@ -8,9 +8,9 @@ import PropTypes from 'prop-types';
 import Modal from 'react-aria-modal';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-import { periodToQuarterMapping } from 'dataMapping/aboutTheData/periodToQuarterMapping';
+import { periodToQuarterMapping } from 'dataMapping/aboutTheData/timeFilters';
 import { showQuarterText } from 'helpers/aboutTheDataHelper';
-import { modalContentMapping } from './dataMapping/modals/modalContentMapping';
+import { modalContentMapping } from './componentMapping/modalContentMapping';
 
 const propTypes = {
     mounted: PropTypes.bool,
