--- src/js/dataMapping/aboutTheData/periodToQuarterMapping.js
@@ -1,19 +1,0 @@
-/**
- * periodToQuarterMapping.js
- * Created by Jonathan Hill 11/20/20
- */
-
-export const periodToQuarterMapping = {
-    1: '1',
-    2: '1',
-    3: '1',
-    4: '2',
-    5: '2',
-    6: '2',
-    7: '3',
-    8: '3',
-    9: '3',
-    10: '4',
-    11: '4',
-    12: '4'
-};
