+++ src/js/dataMapping/aboutTheData/timeFilters.js
@@ -48,3 +48,18 @@ export const periodsPerQuarter = [
         { title: 'Q4 P12', id: '12', className: 'last-period-per-quarter' }
     ]
 ];
+
+export const periodToQuarterMapping = {
+    1: '1',
+    2: '1',
+    3: '1',
+    4: '2',
+    5: '2',
+    6: '2',
+    7: '3',
+    8: '3',
+    9: '3',
+    10: '4',
+    11: '4',
+    12: '4'
+};
